refactor: split battle-turn.ts into focused modules (SRP)
Extract battle state I/O (read/write/delete) into core/battle-state-io.ts
and shared battle setup helpers (move loading, display names, fallback
moves) into core/battle-setup.ts. Both battle-turn CLI and battle-tui
now import from these shared modules, eliminating code duplication.
Add i18n keys for battle messages.
