+++ src/i18n/ko.json
@@ -360,6 +360,20 @@
   "battle.shiny_catch": "\n★ 색이 다른 포켓몬 포획 성공!",
   "battle.shiny_escaped": "\n색이 다른 {pokemon:이/가} 도망쳤다...",
 
+  "battle.gym_challenge": "{leader:이/가} 승부를 걸어왔다!",
+  "battle.send_out": "{leader:은/는} {pokemon:을/를} 내보냈다!",
+  "battle.go": "가라, {pokemon}!",
+  "battle.victory": "{leader}에게 승리했다!",
+  "battle.defeat": "{leader}에게 패배했다...",
+  "battle.select_switch": "교체할 포켓몬을 선택하세요:",
+  "battle.no_switch": "교체할 수 있는 포켓몬이 없다!",
+  "battle.select_next": "다음 포켓몬을 선택하세요:",
+  "battle.all_fainted": "모든 포켓몬이 쓰러졌다...",
+  "battle.surrender_confirm": "정말 항복하시겠습니까?",
+  "battle.press_any_key": "아무 키나 눌러서 나가기",
+  "battle.fainted_label": "[쓰러짐]",
+  "battle.other_session": "다른 세션의 배틀이 진행 중입니다.",
+
   "achievement.unlocked": "🏆 업적 달성: {name}! ",
   "achievement.unlocked_pokemon": "🏆 업적 달성: {name}! {pokemon:을/를} 얻었습니다!",
   "achievement.unlocked_message": "🏆 업적 달성: {name}! {message}",
